move errors to their domains
