+++ src/core/Usecase.ts
@@ -1,0 +1,5 @@
+export class UsecaseError extends Error {
+  constructor(message: string) {
+    super(message);
+  }
+}
